using common fixture
