+++ numbast/src/numbast/tools/tests/additional_include.cuh
@@ -1,3 +1,8 @@
+// clang-format off
+// SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+// clang-format on
+
 #include "bar.cuh"
 
 void __device__ set42(int *arr) { *arr = bar(42); }
