+++ numbast/src/numbast/tools/tests/include/bar.cuh
@@ -1,1 +1,6 @@
+// clang-format off
+// SPDX-FileCopyrightText: Copyright (c) 2025 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+// clang-format on
+
 int __device__ bar(int a) { return a; }
